Added Instagram component in HomeContainer, updated directory structure, added a directory called Common, for navbar, footer, etc
