+++ src/containers/Home.js
@@ -1,13 +1,14 @@
 import React from 'react';
 import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
-import NavBar from '../components/Home/NavBar';
+import NavBar from '../components/Common/NavBar';
 import CarouselBanner from '../components/Home/CarouselBanner';
 import Collection from '../components/Home/Collection';
 import AboutMe from '../components/Home/AboutMe';
 import FeaturedProducts from '../components/Home/FeaturedProducts';
 import Banner from '../components/Home/Banner';
 import BlogHome from '../components/Home/BlogHome';
+import Instagram from '../components/Home/Instagram';
 import '../sass/home.scss';
 
 
@@ -26,6 +27,7 @@ class Home extends React.Component {
                 <FeaturedProducts />
                 <Banner />
                 <BlogHome />
+                <Instagram />
             </div>
         );
     }
