+++ src/sass/home.scss
@@ -94,7 +94,7 @@
     text-align: center;
     padding-bottom: 50px;
   }
-  img{
+  .col-padding-2{
     padding-left: 30px;
     padding-right: 30px;
   }
@@ -142,5 +142,20 @@
     font-family: "Bodoni Sans Text Bold";
     text-decoration: none;
     color: black;
+  }
+}
+
+.Instagram{
+  margin-top: 100px;
+  img{
+    padding-left: 20px;
+    padding-right: 20px;
+  }
+  .box-shadow {
+    background: white;
+    -webkit-box-shadow: -2px -2px 5px 0px rgba(0, 0, 0, 0.1);
+    -moz-box-shadow: -2px -2px 5px 0px rgba(0, 0, 0, 0.1);
+    box-shadow: -2px -2px 5px 0px rgba(0, 0, 0, 0.1);
+    padding: 20px 30px 20px 30px;
   }
 }
